+++ SanghaviMart/src/controller/ReviewV1Controller.h
@@ -8,8 +8,8 @@ namespace sanghavimart::controller {
 class ReviewV1Controller : public drogon::HttpController<ReviewV1Controller> {
   public:
     METHOD_LIST_BEGIN
-    ADD_METHOD_TO(ReviewV1Controller::Add, "/api/v1/reviews", drogon::Post);
-    ADD_METHOD_TO(ReviewV1Controller::List, "/api/v1/products/{id}/reviews", drogon::Get);
+    ADD_METHOD_TO(ReviewV1Controller::Add, "/api/v1/reviews", drogon::Post, "sanghavimart::filter::LoggingFilter");
+    ADD_METHOD_TO(ReviewV1Controller::List, "/api/v1/products/{id}/reviews", drogon::Get, "sanghavimart::filter::LoggingFilter");
     METHOD_LIST_END
 
     void Add(const drogon::HttpRequestPtr& req,